nona_back/requirements.txt
click==8.1.3
Flask==2.2.3
Flask-Cors==3.0.10
urllib3==1.26.15
sqlalchemy==2.0.4
mysql-connector-python==8.0.32
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
PyMySQL==1.0.2
six==1.16.0
typing_extensions==4.5.0
Werkzeug==2.2.3
zipp==3.15.0
firebase-admin==6.1.0
APScheduler==3.10.1
pytest==7.2.2
coverage==7.2.3
pillow==9.0.1
openai
tiktoken
